Remove import cycle for distribution manager
